Refactor base viewer classes to have most of the code be non-Qt-specific
