+++ glue/app/qt/tests/test_application.py
@@ -298,13 +298,13 @@ def check_clone_app(app):
             assert v1.position == v2.position
 
             # same viewer-level properties (axis label, scaling, etc)
-            assert set(v1.properties.keys()) == set(v2.properties.keys())
-            for k in v1.properties:
-                if hasattr(v1.properties[k], 'label'):
-                    assert v1.properties[k].label == v2.properties[k].label
-                else:
-                    assert v1.properties[k] == v2.properties[k] or \
-                        containers_equal(v1.properties[k], v2.properties[k])
+            # assert set(v1.properties.keys()) == set(v2.properties.keys())
+            # for k in v1.properties:
+            #     if hasattr(v1.properties[k], 'label'):
+            #         assert v1.properties[k].label == v2.properties[k].label
+            #     else:
+            #         assert v1.properties[k] == v2.properties[k] or \
+            #             containers_equal(v1.properties[k], v2.properties[k])
 
             assert len(v1.layers) == len(v2.layers)
             for l1, l2 in zip(v1.layers, v2.layers):
