+++ glue/core/application_base.py
@@ -7,16 +7,13 @@ from functools import wraps
 from glue.external.six import string_types
 from glue.core.session import Session
 from glue.core.hub import HubListener
-from glue.core import Data, Subset
-from glue.core import command
+from glue.core import Data
 from glue.core.data_factories import load_data
 from glue.core.data_collection import DataCollection
 from glue.config import settings
-from glue.utils import PropertySetMixin
-from glue.core.command import ApplySubsetState
 
 
-__all__ = ['Application', 'ViewerBase']
+__all__ = ['Application']
 
 
 def catch_error(msg):
@@ -370,177 +367,3 @@ class Application(HubListener):
                 viewer = context.object(v)
                 self.add_widget(viewer, tab=i, hold_position=True)
         return self
-
-
-class ViewerBase(HubListener, PropertySetMixin):
-
-    """ Base class for data viewers in an application """
-
-    # the glue.core.layer_artist.LayerArtistContainer
-    # class/subclass to use
-    _layer_artist_container_cls = None
-
-    def __init__(self, session):
-
-        HubListener.__init__(self)
-        PropertySetMixin.__init__(self)
-
-        self._session = session
-        self._data = session.data_collection
-        self._hub = None
-        self._layer_artist_container = self._layer_artist_container_cls()
-
-    def close(self):
-        self._layer_artist_container.clear()
-
-    def register_to_hub(self, hub):
-        self._hub = hub
-
-    def unregister(self, hub):
-        """ Abstract method to unsubscribe from messages """
-        raise NotImplementedError
-
-    def request_add_layer(self, layer):
-        """ Issue a command to add a layer """
-        cmd = command.AddLayer(layer=layer, viewer=self)
-        self._session.command_stack.do(cmd)
-
-    def add_layer(self, layer):
-        if isinstance(layer, Data):
-            self.add_data(layer)
-        elif isinstance(layer, Subset):
-            self.add_subset(layer)
-        # else: SubsetGroup
-
-    def add_data(self, data):
-        """ Add a data instance to the viewer
-
-        This must be overridden by a subclass
-
-        Parameters
-        ----------
-        data : :class:`~glue.core.data.Data`
-            Data object to add.
-        """
-        raise NotImplementedError
-
-    def add_subset(self, subset):
-        """ Add a subset to the viewer
-
-        This must be overridden by a subclass
-
-        Parameters
-        ----------
-        subset : :class:`~glue.core.subset.Subset`
-            Subset instance to add.
-        """
-        raise NotImplementedError
-
-    def apply_roi(self, roi):
-        """ Apply an ROI to the client
-
-        Parameters
-        ----------
-        roi : :class:`~glue.core.roi.Roi`
-            The ROI to apply.
-        """
-        raise NotImplementedError
-
-    @property
-    def session(self):
-        return self._session
-
-    def apply_subset_state(self, subset_state, use_current=False):
-        cmd = ApplySubsetState(data_collection=self._data,
-                               subset_state=subset_state,
-                               use_current=use_current)
-        self._session.command_stack.do(cmd)
-
-    def layer_view(self):
-        raise NotImplementedError()
-
-    def options_widget(self):
-        raise NotImplementedError()
-
-    def move(self, x=None, y=None):
-        """ Reposition a viewer within the application.
-
-        x : int, optional
-            Offset of viewer's left edge from the left edge of the parent
-            window.
-        y : int, optional
-            Offset of the viewer's top edge from the top edge of the parent
-            window.
-        """
-        raise NotImplementedError()
-
-    @property
-    def position(self):
-        """
-        Return the location of the viewer as a tuple of ``(x, y)``
-        """
-        raise NotImplementedError()
-
-    @property
-    def viewer_size(self):
-        """
-        Return the size of the viewer as a tuple of ``(width, height)``
-        """
-        raise NotImplementedError()
-
-    @viewer_size.setter
-    def viewer_size(self, value):
-        """ Resize the width and/or height of the viewer
-
-        Parameters
-        ----------
-        value : tuple of int
-            The width and height of the viewer.
-        width : int, optional
-            New width.
-        height : int, optional
-            New height.
-        """
-        raise NotImplementedError()
-
-    def restore_layers(self, rec, context):
-        """
-        Given a list of glue-serialized layers, restore them
-        to the viewer
-        """
-        # if this viewer manages a client, rely on it to restore layers
-        if hasattr(self, 'client'):
-            return self.client.restore_layers(rec, context)
-        raise NotImplementedError()
-
-    @property
-    def layers(self):
-        """Return a tuple of layers in this viewer.
-
-        A layer is a visual representation of a dataset or subset within
-        the viewer"""
-        return tuple(self._layer_artist_container)
-
-    def __gluestate__(self, context):
-        return dict(session=context.id(self._session),
-                    size=self.viewer_size,
-                    pos=self.position,
-                    properties=dict((k, context.id(v))
-                                    for k, v in self.properties.items()),
-                    layers=list(map(context.do, self.layers))
-                    )
-
-    @classmethod
-    def __setgluestate__(cls, rec, context):
-        session = context.object(rec['session'])
-        result = cls(session)
-        result.register_to_hub(session.hub)
-        result.viewer_size = rec['size']
-        x, y = rec['pos']
-        result.move(x=x, y=y)
-
-        prop = dict((k, context.object(v)) for
-                    k, v in rec['properties'].items())
-        result.restore_layers(rec['layers'], context)
-        result.properties = prop
-        return result
