+++ glue/viewers/common/qt/tests/test_data_viewer.py
@@ -29,7 +29,7 @@ class BaseTestDataViewer(object):
 
         w = self.widget_cls(session)
         w.register_to_hub(hub)
-        with patch.object(DataViewer, 'unregister') as unregister:
+        with patch.object(w, 'unregister') as unregister:
             w.close()
         unregister.assert_called_once_with(hub)
